// src/app/registration/model/registration.model.ts
export class Registration {
    exhibitorName: string = '';
    address: string = '';
    city: string = '';
    state: string = '';
    country: string = '';
    telephone: string = '';
    email: string = '';
    mobile: string = '';
    name: string = '';
    password: string = '';
    confirmPassword: string = '';
}